conda_requirements.txt
###
# Intent Gateway is a natural language processing (NLP) framework to conduct model development and
# model deployment for text classification.
# Copyright (C) 2018-2019  Asurion, LLC
#
# Intent Gateway is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# Intent Gateway is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with Intent Gateway.  If not, see <https://www.gnu.org/licenses/>.
###

# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

absl-py=0.7.1
asn1crypto=0.24.0
astor=0.7.1
attrs=19.1.0
backports=1
backports.functools_lru_cache=1.5
backports.weakref=1.0.post1
backports_abc=0.5
bcrypt=3.1.6
blas=1.1
blinker=1.4
boto=2.49.0
boto3=1.7.30
botocore=1.10.30
bz2file=0.98
c-ares=1.15.0
ca-certificates=2018.4.16
certifi=2018.4.16
cffi=1.12.3
chardet=3.0.4
click=6.7
cryptography=2.3.1
cryptography-vectors=2.3.1
cycler=0.10.0
cymem=2.0.2
cython-blis=0.2.4
cytoolz=0.8.2
dbus=1.11.0
dill=0.2.8.2
docutils=0.14
enum34=1.1.6
expat=2.2.5
fabric=2.4.0
flask=1.0.2
flask-cors=3.0.4
fontconfig=2.13.0
freetype=2.8.1
funcsigs=1.0.2
functools32=3.2.3.2
futures=3.2.0
fuzzywuzzy=0.16.0
gast=0.2.2
gensim=3.4.0
glib=2.53.6
grpcio=1.14.1
gst-plugins-base=1.12.4
gstreamer=1.12.4
gunicorn=19.8.1
h5py=2.9.0
hdf5=1.10.4
icu=58.2
idna=2.6
intel-openmp=2019.3
invoke=1.2.0
ipaddress=1.0.22
itsdangerous=1.1.0
jmespath=0.9.3
joblib=0.11
jpeg=9b
jpype1=0.6.3
jsonschema=3.0.0a3
keras=2.2.2
keras-applications=1.0.4
keras-preprocessing=1.0.2
kiwisolver=1.0.1
libblas=3.8.0
libcblas=3.8.0
libedit=3.1.20170329
libffi=3.2.1
libgcc=7.2.0
libgcc-ng=8.2.0
libgfortran=3.0.0
libgfortran-ng=7.3.0
libgpuarray=0.7.6
libiconv=1.15
liblapack=3.8.0
libopenblas=0.3.3
libpng=1.6.34
libprotobuf=3.7.1
libstdcxx-ng=8.2.0
libuuid=2.32.1
libxcb=1.13
libxml2=2.9.8
linecache2=1.0.0
mako=1.0.7
markdown=2.6.11
markupsafe=1
matplotlib=2.2.2
mkl=2019.3
mock=3.0.5
msgpack-numpy=0.4.4.2
msgpack-python=0.4.8
murmurhash=0.28.0
ncurses=6.1
nltk=3.2.5
numpy=1.16.3
numpy-base=1.16.3
oauthlib=2.0.7
openblas=0.2.20
openssl=1.0.2o
pandas=0.23.0
paramiko=2.4.2
pathlib=1.0.1
pcre=8.41
pip=19.1.1
plac=0.9.6
preshed=2.0.1
prettytable=0.7.2
protobuf=3.7.1
pthread-stubs=0.4
pycosat=0.6.3
pycparser=2.18
pygpu=0.7.6
pyjwt=1.6.3
pynacl=1.3.0
pyopenssl=17.5.0
pyparsing=2.2.0
pyqt=5.6.0
pyrsistent=0.15.1
pysocks=1.6.8
python=2.7.14
python-crfsuite=0.9.5
python-dateutil=2.7.3
python-levenshtein=0.12.0
pytz=2018.4
pyyaml=3.12
qt=5.6.2
readline=7
regex=2019.04.14
requests=2.18.4
requests-oauthlib=0.8.0
ruamel_yaml=0.15.35
s3transfer=0.1.13
scikit-learn=0.20.3
scipy=1.1.0
setuptools=39.2.0
singledispatch=3.4.0.3
sip=4.18
six=1.11.0
smart_open=1.5.7
spacy=2.1.3
sqlite=3.23.1
srsly=0.0.5
subprocess32=3.5.1
tensorboard=1.10.0
tensorflow=1.10.0
termcolor=1.1.0
theano=1.0.3
thinc=7.0.4
tk=8.6.7
toolz=0.9.0
tornado=5.0.2
tqdm=4.23.4
traceback2=1.4.0
trollius=2.2
twython=3.7.0
ujson=1.35
unittest2=1.1.0
urllib3=1.22
wasabi=0.2.2
werkzeug=0.14.1
wheel=0.31.0
wrapt=1.10.11
xgboost=0.71
xorg-libxau=1.0.8
xorg-libxdmcp=1.1.2
xz=5.2.3
yaml=0.1.7
zlib=1.2.11